Split LV2 ports to a new header file...
... Also split out a small header for the RAII utilities
