+++ src/effects/lv2/LV2Effect.h
@@ -17,7 +17,6 @@
 
 class wxArrayString;
 
-#include <memory>
 #include <thread>
 #include <vector>
 
@@ -34,12 +33,11 @@ class wxArrayString;
 #include "lv2_external_ui.h"
 
 #include "LV2FeaturesList.h"
+#include "LV2Ports.h"
 #include "../../ShuttleGui.h"
 #include "SampleFormat.h"
 
 #include "NativeWindow.h"
-
-#include "zix/ring.h"
 
 #include <unordered_map>
 
@@ -64,174 +62,7 @@ class NumericTextCtrl;
 
 // DECLARE_LOCAL_EVENT_TYPE(EVT_SIZEWINDOW, -1);
 
-//! Immutable description of an LV2 port
-class LV2Port {
-public:
-   LV2Port(const LilvPort *port, int index, bool isInput,
-      const wxString &symbol, const wxString &name,
-      const TranslatableString &group
-   )  : mPort(port), mIndex(index), mIsInput(isInput)
-      , mSymbol(symbol), mName(name), mGroup(group)
-   {}
-   const LilvPort *const mPort;
-   const uint32_t mIndex;
-   const bool mIsInput;
-   const wxString mSymbol;
-   const wxString mName;
-   const TranslatableString mGroup;
-};
-
-//! Immutable description of an LV2 Audio port
-class LV2AudioPort final : public LV2Port {
-public:
-   using LV2Port::LV2Port;
-};
-using LV2AudioPortPtr = std::shared_ptr<LV2AudioPort>;
-using LV2AudioPortArray = std::vector<LV2AudioPortPtr>;
-
-//! Immutable description of an LV2 Atom port
-class LV2AtomPort final : public LV2Port {
-public:
-   LV2AtomPort(const LilvPort *port, int index, bool isInput,
-      const wxString & symbol, const wxString & name,
-      const TranslatableString & group,
-      uint32_t minimumSize, bool isMidi, bool wantsPosition
-   )  : LV2Port{ port, index, isInput, symbol, name, group }
-      , mMinimumSize{ minimumSize }
-      , mIsMidi{ isMidi }
-      , mWantsPosition{ wantsPosition }
-   {}
-   const uint32_t mMinimumSize;
-   const bool mIsMidi;
-   const bool mWantsPosition;
-};
-using LV2AtomPortPtr = std::shared_ptr<LV2AtomPort>;
-using LV2AtomPortArray = std::vector<LV2AtomPortPtr>;
-
-//! State of an instance of an LV2 Atom port
-struct LV2AtomPortState final {
-   //! @pre `pPort != nullptr`
-   explicit LV2AtomPortState(LV2AtomPortPtr pPort)
-      : mpPort{ move(pPort) }
-      , mRing{ zix_ring_new(mpPort->mMinimumSize) }
-      , mBuffer( mpPort->mMinimumSize )
-   {
-      assert(mpPort);
-      zix_ring_mlock(mRing.get());
-   }
-   const LV2AtomPortPtr mpPort;
-   const Lilv_ptr<ZixRing, zix_ring_free> mRing;
-   std::vector<uint8_t> mBuffer;
-};
-using LV2AtomPortStatePtr = std::shared_ptr<LV2AtomPortState>;
-using LV2AtomPortStateArray = std::vector<LV2AtomPortStatePtr>;
-
-//! Immutable description of an LV2 CV port (control data signal at sample rate)
-class LV2CVPort final : public LV2Port {
-public:
-   LV2CVPort(const LilvPort *port, int index, bool isInput,
-      const wxString & symbol, const wxString & name,
-      const TranslatableString & group,
-      float min, float max, float def, bool hasLo, bool hasHi
-   )  : LV2Port(port, index, isInput, symbol, name, group)
-      , mMin{ min }, mMax{ max }, mDef{ def }, mHasLo{ hasLo }, mHasHi{ hasHi }
-   {}
-   const float mMin;
-   const float mMax;
-   const float mDef;
-   const bool mHasLo;
-   const bool mHasHi;
-};
-using LV2CVPortPtr = std::shared_ptr<LV2CVPort>;
-using LV2CVPortArray = std::vector<LV2CVPortPtr>;
-
-//! State of an instance of an LV2 CV port
-struct LV2CVPortState final {
-   //! @pre `pPort != nullptr`
-   explicit LV2CVPortState(LV2CVPortPtr pPort) : mpPort{ move(pPort) } {
-      assert(mpPort);
-   }
-   const LV2CVPortPtr mpPort;
-   Floats mBuffer;
-};
-//! No need yet for extra indirection
-using LV2CVPortStateArray = std::vector<LV2CVPortState>;
-
 class LV2EffectMeter;
-
-//! Immutable description of an LV2 control port
-class LV2ControlPort final : public LV2Port
-{
-public:
-   LV2ControlPort(const LilvPort *port, int index, bool isInput,
-      const wxString & symbol, const wxString & name,
-      const TranslatableString & group,
-      std::vector<double> scaleValues, wxArrayString scaleLabels,
-      const wxString &units,
-      float min, float max, float def, bool hasLo, bool hasHi,
-      bool toggle, bool enumeration, bool integer, bool sampleRate,
-      bool trigger, bool logarithmic
-   )  : LV2Port{ port, index, isInput, symbol, name, group }
-      , mScaleValues{ move(scaleValues) }
-      , mScaleLabels{ std::move(scaleLabels) }
-      , mUnits{ units }
-      , mMin{ min }, mMax{ max }, mDef{ def }
-      , mHasLo{ hasLo }, mHasHi{ hasHi }
-      , mToggle{ toggle }, mEnumeration{ enumeration }, mInteger{ integer }
-      , mSampleRate{ sampleRate }
-      , mTrigger{ trigger }, mLogarithmic{ logarithmic }
-   {}
- 
-   // ScalePoints
-   const std::vector<double> mScaleValues;
-   const wxArrayString mScaleLabels;
-
-   const wxString mUnits;
-   const float mMin;
-   const float mMax;
-   const float mDef;
-   const bool mHasLo;
-   const bool mHasHi;
-   const bool mToggle;
-   const bool mEnumeration;
-   const bool mInteger;
-   const bool mSampleRate;
-   const bool mTrigger;
-   const bool mLogarithmic;
-
-   //! Map a real number to one of the scale points
-   size_t Discretize(float value) const {
-      auto s = mScaleValues.size();
-      for (; s > 0 && --s;)
-         if (value >= mScaleValues[s])
-            break;
-      return s;
-   }
-};
-using LV2ControlPortPtr = std::shared_ptr<LV2ControlPort>;
-using LV2ControlPortArray = std::vector<LV2ControlPortPtr>;
-
-//! State of an instance of an LV2 Control port
-struct LV2ControlPortState final {
-   //! @pre `pPort != nullptr`
-   explicit LV2ControlPortState(LV2ControlPortPtr pPort)
-      : mpPort{ move(pPort) }
-   {
-      assert(mpPort);
-   }
-   const LV2ControlPortPtr mpPort;
-   //! Value of mTmp last seen by idle-time updater
-   float mLst{ 0.0 };
-   //! Value of UI control, as scaled by sample rate if that is required
-   float mTmp{ 0.0 };
-   //! Lower bound, as scaled by sample rate if that is required
-   float mLo{ 0.0 };
-   //! Upper bound, as scaled by sample rate if that is required
-   float mHi{ 0.0 };
-};
-//! No need yet for extra indirection
-using LV2ControlPortStateArray = std::vector<LV2ControlPortState>;
-
 class LV2Wrapper;
 
 struct LV2EffectSettings {
@@ -536,23 +367,6 @@ private:
    DECLARE_EVENT_TABLE()
 
    friend class LV2Wrapper;
-};
-
-//! Use when lilv.h comments "must not be freed" or we use the node elsewhere,
-//! or the node pointer is from iterating a LilvNodes collection
-inline wxString LilvString(const LilvNode *node)
-{
-   return wxString::FromUTF8(lilv_node_as_string(node));
-};
-
-//! Use when lilv.h comments "Returned value must be freed by the caller."
-//! We free it in this function.
-//! Name suggests C++ move semantics applied to `node`, but only C types used
-inline wxString LilvStringMove(LilvNode *node)
-{
-   LilvNodePtr temp{ node };
-   wxString str = LilvString(node);
-   return str;
 };
 
 class LV2Wrapper final
