+++ src/effects/lv2/LV2Symbols.h
@@ -5,7 +5,7 @@
   @file LV2Symbols.h
 
   @brief Declare URI identifiers used in calls to the lv2 library and a mapping
-  of them to integers; also a smart pointer template for managing lv2 resources
+  of them to integers
 
   Paul Licameli split from LoadLV2.h
 
@@ -16,19 +16,8 @@
 
 #include <vector>
 #include "lilv/lilv.h"
+#include "LV2Utils.h"
 #include "MemoryX.h"
-
-//! Generate deleter classes for smart pointers to lv2 resources
-template<typename Type, void (*f)(Type*)>
-struct Lilv_deleter { void operator()(Type *p) noexcept { f(p); } };
-
-//! Generate classes of smart pointers to lv2 resources
-template<typename Type, void (*f)(Type*)>
-using Lilv_ptr = std::unique_ptr<Type, Lilv_deleter<Type, f>>;
-
-//! A smart pointer to LilvNode, which is essentially a variant type much used
-//! in the lv2 API
-using LilvNodePtr = Lilv_ptr<LilvNode, lilv_node_free>;
 
 #if !defined(LV2_CORE__MIDIPlugin)
 #define LV2_CORE__MIDIPlugin   LV2_CORE_PREFIX "MIDIPlugin"
